Boss-shooting and AnimationFactory implemented

+++ src/main/java/projectiles/BombProjectile.java
@@ -34,12 +34,12 @@ public class BombProjectile extends Entity implements Projectile {
     }
 
     @Override
-    public int getXSpeed() {
+    public float getXSpeed() {
         return this.xSpeed;
     }
 
     @Override
-    public int getYSpeed() {
+    public float getYSpeed() {
         return this.ySpeed;
     }
