+++ src/main/java/projectiles/FlameProjectile.java
@@ -37,12 +37,12 @@ public class FlameProjectile extends Entity implements Projectile {
    }
 
    @Override
-   public int getXSpeed() {
+   public float getXSpeed() {
       return this.xSpeed;
    }
 
    @Override
-   public int getYSpeed() {
+   public float getYSpeed() {
       return this.ySpeed;
    }
