+++ src/main/java/projectiles/OctaProjectile.java
@@ -36,12 +36,12 @@ public class OctaProjectile extends Entity implements Projectile {
     }
 
     @Override
-    public int getXSpeed() {
+    public float getXSpeed() {
         return this.xSpeed;
     }
 
     @Override
-    public int getYSpeed() {
+    public float getYSpeed() {
         return this.ySpeed;
     }
