+++ src/main/java/projectiles/PlayerProjectile.java
@@ -42,12 +42,12 @@ public class PlayerProjectile extends Entity implements Projectile {
     }
 
     @Override
-    public int getXSpeed() {
+    public float getXSpeed() {
         return this.xSpeed;
     }
 
     @Override
-    public int getYSpeed() {
+    public float getYSpeed() {
         return this.ySpeed;
     }
